byungmin : 에러 공통으로 예외 처리

+++ src/main/java/com/hanghae/mini2/riceFriend/controller/MeetingRestController.java
@@ -4,6 +4,7 @@ import com.hanghae.mini2.riceFriend.config.auth.PrincipalDetails;
 import com.hanghae.mini2.riceFriend.dto.request.MeetingRequestDto;
 import com.hanghae.mini2.riceFriend.dto.response.MeetingDetailResponseDto;
 import com.hanghae.mini2.riceFriend.dto.response.MeetingResonseDto;
+import com.hanghae.mini2.riceFriend.handler.ex.LoginUserNotFoundException;
 import com.hanghae.mini2.riceFriend.handler.ex.MeetingRequestException;
 import com.hanghae.mini2.riceFriend.service.MeetingService;
 import io.swagger.annotations.Api;
@@ -50,47 +51,14 @@ public class MeetingRestController {
         return meetingDetailResponseDto;
     }
 
-    // (테스트중...파일업로드)
-//    @PostMapping("/api/meeting/test")
-//    @ApiOperation(value = "맛집모임 정보 등록.", notes = "맛집모임 정보를 입력받아 등록한다.")
-//    public HashMap<String, Object> createMeeting(Long userId,
-//                                                 @RequestParam("image") MultipartFile multipartFile) throws IOException {
-//
-////        if (userDetails == null)
-////            throw new MeetingRequestException("로그인 한 사용자만 작성할 수 있습니다.");
-//
-//        /////////////////////////////테스트 데이터영역////////////////////////////////
-//        userId = 5L;
-//        MeetingRequestDto meetingRequestDto = MeetingRequestDto.builder()
-//                .restaurantName("음식점이름_test5")
-//                .imgUrl("음식점url_test5")
-//                .locationId(11L)
-//                .meetingTitle("모임날짜_test5")
-//                .content("모임내용_test5")
-//                .meetingDate(LocalDateTime.now())
-//                .limitMember(2)
-//                .build();
-//
-//        meetingService.createMeeting(meetingRequestDto, multipartFile, userId);
-//        /////////////////////////////////////////////////////////////////////////
-//
-//        HashMap<String, Object> result = new HashMap<>();
-//
-////        meetingService.createMeeting(requestDto, multipartFile, userDetails.getUser());
-//        result.put("result", "true");
-//
-//        return result;
-//    }
-
     @PostMapping("/api/meeting")
     @ApiOperation(value = "맛집모임 정보 등록.", notes = "맛집모임 정보를 입력받아 등록한다.")
     public HashMap<String, Object> createMeeting(@RequestPart(value = "data") @Valid MeetingRequestDto requestDto,
                                                  @RequestPart(value = "image", required = false) MultipartFile multipartFile,
                                                  @AuthenticationPrincipal PrincipalDetails userDetails) throws IOException {
 
-        //System.out.println("userDetails = " + userDetails);
         if (userDetails == null)
-            throw new MeetingRequestException("로그인 한 사용자만 작성할 수 있습니다.");
+            throw new LoginUserNotFoundException("로그인한 유저 정보가 없습니다.");
 
         HashMap<String, Object> result = new HashMap<>();
 
@@ -100,42 +68,15 @@ public class MeetingRestController {
         return result;
     }
 
-    // (테스트중...파일업로드)
-//    @PutMapping("/api/meeting/test2")
-//    @ApiOperation(value = "맛집모임 정보 수정.", notes = "맛집모임 정보를 수정한다.")
-//    public HashMap<String, Object> updateMeeting(Long meeting_id,
-//                                                 @RequestParam("image") MultipartFile multipartFile,
-//                                                 Long userId) throws IOException{
-//
-//        HashMap<String, Object> result = new HashMap<>();
-//
-//        /////////////////////////////테스트 데이터영역////////////////////////////////
-//        meeting_id = 1L;
-//        userId = 5L;
-//        MeetingRequestDto meetingRequestDto = MeetingRequestDto.builder()
-//                .restaurantName("음식점이름_test3")
-//                .restaurantUrl("음식점url_test3")
-//                .locationId(11L)
-//                .meetingTitle("모임날짜_test3")
-//                .content("모임내용_test3")
-//                .meetingDate(LocalDateTime.now())
-//                .limitMember(3)
-//                .build();
-//
-//        meetingService.updateMeeting(meeting_id, meetingRequestDto, multipartFile, userId);
-//        /////////////////////////////////////////////////////////////////////////
-//
-//        result.put("result", "true");
-//
-//        return result;
-//    }
-
     @PutMapping("/api/meeting/{meeting_id}")
     @ApiOperation(value = "맛집모임 정보 수정.", notes = "맛집모임 정보를 수정한다.")
     public HashMap<String, Object> updateMeeting(@PathVariable Long meeting_id,
-                                                 @RequestPart(value = "data") MeetingRequestDto requestDto,
+                                                 @RequestPart(value = "data") @Valid MeetingRequestDto requestDto,
                                                  @RequestPart(value = "image", required = false) MultipartFile multipartFile,
                                                  @AuthenticationPrincipal PrincipalDetails userDetails) throws IOException{
+
+        if (userDetails == null)
+            throw new LoginUserNotFoundException("로그인한 유저 정보가 없습니다.");
 
         HashMap<String, Object> result = new HashMap<>();
 
@@ -162,6 +103,10 @@ public class MeetingRestController {
     @ApiOperation(value = "맛집모임 정보 삭제.", notes = "맛집모임 정보를 삭제한다.")
     public HashMap<String, Object> deleteMeeting(@PathVariable Long meeting_id,
                                                  @AuthenticationPrincipal PrincipalDetails userDetails) {
+
+        if (userDetails == null)
+            throw new LoginUserNotFoundException("로그인한 유저 정보가 없습니다.");
+
         HashMap<String, Object> result = new HashMap<>();
 
         meetingService.deleteMeeting(meeting_id, userDetails.getUser().getId());
@@ -174,6 +119,10 @@ public class MeetingRestController {
     @ApiOperation(value = "모임 참여.", notes = "유저가 선택한 모임에 참여한다.")
     public HashMap<String, Object> createMeetingUser(@PathVariable Long meeting_id,
                                                      @AuthenticationPrincipal PrincipalDetails userDetails) {
+
+        if (userDetails == null)
+            throw new LoginUserNotFoundException("로그인한 유저 정보가 없습니다.");
+
         HashMap<String, Object> result = new HashMap<>();
 
         meetingService.createMeetingUser(meeting_id, userDetails.getUser().getId());
@@ -186,6 +135,10 @@ public class MeetingRestController {
     @ApiOperation(value = "모임 탈퇴.", notes = "유저가 참여한 모임에 탈퇴한다.")
     public HashMap<String, Object> deleteMeetingUser(@PathVariable Long meeting_id,
                                                      @AuthenticationPrincipal PrincipalDetails userDetails) {
+
+        if (userDetails == null)
+            throw new LoginUserNotFoundException("로그인한 유저 정보가 없습니다.");
+
         HashMap<String, Object> result = new HashMap<>();
 
         meetingService.deleteMeetingUser(meeting_id, userDetails.getUser().getId());
